+++ frontend/src/js/pages/content/ActionPanel.tsx
