+++ frontend/src/js/pages/content/DeleteSelected.tsx
