+++ frontend/src/js/pages/content/Display.tsx
