+++ frontend/src/js/pages/content/SearchBar.tsx
