+++ frontend/src/js/pages/content/index.tsx
