+++ frontend/src/js/pages/home/Icons.tsx
