+++ frontend/src/js/pages/metadata/icon.png
